receiver/requirements.txt
aiosmtpd==1.4.4.post2
pycryptodome==3.18.0